fix: serialize uint64_t and int64_t

+++ request.cpp
@@ -1,9 +1,9 @@
 #include "request.hpp"
 
-Request::Request(std::vector<float> vec){
+Request::Request(std::vector<double> vec){
     this->vec = vec;
     meta.key_fields.push_back("vec");
-    meta.type_fields["vec"] = ARRAY + FLOAT;
+    meta.type_fields["vec"] = ARRAY + DOUBLE;
     meta.ptr_fields["vec"] =  &this->vec;
     std::cout << "vec size: " << vec.size() << std::endl;
 }
@@ -11,7 +11,7 @@ Request::Request(std::vector<float> vec){
 
 Request::Request(){
     meta.key_fields.push_back("vec");
-    meta.type_fields["vec"] = ARRAY + FLOAT;
+    meta.type_fields["vec"] = ARRAY + DOUBLE;
     meta.ptr_fields["vec"] =  &this->vec;
 }
